Created base shared class for GameModifiers
