+++ TownOfUs/Modifiers/Game/AllianceGameModifier.cs
@@ -1,6 +1,5 @@
 ﻿using AmongUs.GameOptions;
 using MiraAPI.Modifiers;
-using MiraAPI.Modifiers.Types;
 using MiraAPI.PluginLoading;
 using TownOfUs.Modifiers.Neutral;
 using TownOfUs.Roles.Other;
@@ -8,30 +7,20 @@ using TownOfUs.Roles.Other;
 namespace TownOfUs.Modifiers.Game;
 
 [MiraIgnore]
-public abstract class AllianceGameModifier : GameModifier
+public abstract class AllianceGameModifier : TouBaseGameModifier
 {
-    public virtual string LocaleKey => "KEY_MISS";
-    public virtual string IntroInfo => $"{TouLocale.Get("Alliance")}: {ModifierName}";
+    public override string IntroInfo => $"{TouLocale.Get("Alliance")}: {ModifierName}";
     public virtual string Symbol => "?";
-    public virtual float IntroSize => 4f;
-    public virtual int CustomAmount => GetAmountPerGame();
-    public virtual int CustomChance => GetAssignmentChance();
+    public override float IntroSize => 4f;
     public virtual bool DoesTasks => true;
     public virtual bool GetsPunished => true;
     public virtual bool CrewContinuesGame => true;
-    public virtual ModifierFaction FactionType => ModifierFaction.Alliance;
+    public override ModifierFaction FactionType => ModifierFaction.Alliance;
     // Set to Crewmate, Impostor or Neutral
     public virtual AlliedFaction TrueFactionType => AlliedFaction.Other;
     public virtual bool CountTowardsTrueFaction => false;
     // Only used when the TrueFactionType is set to RoleSpecific
     public virtual RoleBehaviour UnderlyingRole => RoleManager.Instance.GetRole(RoleTypes.Crewmate);
-
-    public override bool HideOnUi => false;
-
-    public override int GetAmountPerGame()
-    {
-        return 1;
-    }
 
     public override bool IsModifierValidOn(RoleBehaviour role)
     {
